Fix filepath & warning

+++ src/main/scala/com/github/frankivo/Graphs.scala
@@ -33,7 +33,7 @@ class Graphs extends Actor {
   def mapData(data: Seq[CovidRecord]): DefaultCategoryDataset = {
     val dataset = new DefaultCategoryDataset
 
-    data.foreach(s => dataset.setValue(s.count, "Cases", s.date.getDayOfMonth))
+    data.foreach(s => dataset.setValue(s.count.toDouble, "Cases", s.date.getDayOfMonth))
     dataset
   }
